Remove invalid asm clobbers
An asm clobber is not allowed to specify a register used as an input or
output.  In the case of the debug asms which use "a" for input, the
accumulator is never actually clobbered by the asm body, so just remove
it from the clobber list.  In the case of fc_nyblswap, the accumulator
is used as both input and output, so declare it as such instead of as
a clobber.