Fix fc_nyblswap on LLVM
With a constraint "a", %0 will expand to "A" in the asm, meaning
"ld%0" will become "ldA" without any address.  But since we have
constrained the operand to already be in A, there is no need to add
any load or store instructions anyway, so just remove them.

Also the global variable "swp" is not needed either, so remove that as well
when using LLVM.